Macro expander/preprocessor.
The macro language is processed right after lexing, before parsing, and
is token-based. It is inspired by Rust's macro facility.

Macros define transformations from token lists to token lists. Arguments
are token lists as well. The only restriction on the macro body and on
arguments is that they be balanced in parentheses/brackets/braces (this
is how the macro processor splits args and delimits macro bodies
properly). A macro can define multiple 'match arms', and the first
matching arm is used. Substitutions can be recursive or corecursive --
the output of a macro is re-scanned until no further macro invocations
remain. The argument list functionality supports sublists (e.g.
my_macro!([x, y, z], 1)) and variadic args to allow a sort of
Scheme-like "first/rest" pattern (see test) in lieu of true loops.

This is somewhat primitive but it will hopefully be sufficient to build
some interesting things.