Support for multiple writers to ports, chans, and regs.
Multiple write-points may exist for these primitives as long as (i) they
are scheduled into the same pipeline stage, and (ii) they have disjoint
predicates. The lowering transform checks both of these conditions, and
then rewrites the set of writes to a single write with a MUX-tree on the
input. (Array writes are exempt from this: each write site becomes a
separate write port. To do better would actually require solving a
set-cover optimization problem with the predicates to find the minimum
number of write points.)