Support for "kill writes" to writable objects (ports/chans/regs).
A "killing write" is a write that occurs in the same stage as, and is
dominated (must be preceded) by, a killyounger. Such a write is able to
override a write to the same resource in an earlier stage because the
killyounger kills the earlier stage's write. Thus, killing writes are
exempt from the same-stage constraint. This change properly wires them
up (via stage-crossing links using RestartValue/RestartValueSrc IR ops)
so that they work.